Fix expressions with multiple operators. (#12)
* Fix expressions with multiple operators.

Expressions such as ``1 +2;`` or ``300*200`` work, but expressions which have multiple operations do not work. For example: ``1+2+3+4`` does not work, and has to be put in a chain of parentheses to get it to work. The other valid expressions are getting evaluated to the value of the first operator, for instance, ``1*2*3`` becomes 2, by evaluating 1*2 and so on.
This can be solved by adding a loop, converting the first two expressions into ``lhs`` and continuing until there are no more ``pair``s.

* Added tests

* Fixed test

Fixed mistake in making assertion of ``1 + 2 +3 == 2+2+3``

* Style correction.

* Update parser.rs

Yet another attempt to make rustfmt happy.